ci: enabled daily dependabot updates for github actions
Introduced a Dependabot configuration to automatically monitor and
propose updates for GitHub Actions and Python dependencies on a daily
basis. Scheduled updates are set for 10:00 and 11:00 CET, respectively,
to ensure consistent dependency management and reduce potential
integration issues. Specific labels are assigned to each ecosystem for
easy categorization of update PRs.