docs: add installation guide for mdsanima-amarok
Enhanced the README with a new section detailing the installation
process for _mdsanima-amarok_ program, including commands for build
and install operations.

Instructions cover how to run the program globally by placing it
in `/usr/local/bin`. Additional changes outline the creation of
`mdsanima.h` in `/usr/local/include` and `libmdsanima.a` in
`/usr/local/lib` to support project integration.