refactor: use shared library in fantastic example
Moved the printing logic from the main executable to a new shared
library in the fantastic example, streamlining the `main.cc` file.
This change encourages modularity and demonstrates shared library
usage in the project. By extracting logic into `lib.cc` and including
it as a dynamic library, we set a precedent for future development
practices and improved the project structure for easier maintenance
and enhancement.

- Introduced `fantastic` as a shared library.
- Updated `main.cc` to utilize the new library for printing messages.
- Added `lib.cc` and `lib.hh` to implement the message printing
functionality in a separate module.

This modification not only makes the code more organized but also
provides a practical example of how to create and link shared
libraries within a CMake project