style: refine documentation
Updated `commitlint` and labels configuration comments for better
clarity and added a new general label category in the `labels.yml`
file.

This change aims to make the purpose and usage of configurations
easier to understand, especially for new contributors. It streamlines
the documentation within the commitlint config file, ensuring that
the guidelines for commit messages are clearer. Additionally, the
introduction of a _General labels_ section in the labels configuration
will help in categorizing issues and pull requests more systematically.